feature(refactor): Refactoring all packages
dispatch classes into core and demo packages

- `core` will host framework classes
- `demo` will host samples code.

fix #23
