+++ src/main/java/fr/snapgames/fromclasstogame/core/entity/TextObject.java
@@ -1,4 +1,4 @@
-package fr.snapgames.fromclasstogame;
+package fr.snapgames.fromclasstogame.core.entity;
 
 import java.awt.*;
