+++ src/main/java/fr/snapgames/fromclasstogame/core/exceptions/io/UnknownResource.java
@@ -1,4 +1,4 @@
-package fr.snapgames.fromclasstogame.io.exception;
+package fr.snapgames.fromclasstogame.core.exceptions.io;
 
 public class UnknownResource extends Exception {
     public UnknownResource(String message,Exception e) {
