+++ src/main/resources/config.properties
@@ -10,6 +10,6 @@ game.setup.scale=2.0
 # Game frame per seconds
 game.setup.fps=60
 # Game scene list (comma sepearets all entries and one entry is 'name:class.with.path')
-game.setup.scenes=demo:fr.snapgames.fromclasstogame.scenes.DemoScene
+game.setup.scenes=demo:fr.snapgames.fromclasstogame.demo.scenes.DemoScene
 # the default scene to be loaded
 game.setup.scene.default=demo
